[UT] Move PartialRuntimeFilterMerger tests to the full-pipeline test binary
The merger tests call PartialRuntimeFilterMerger::add_partial_filters, which is
defined in the pipeline runtime sources (PIPELINE_RUNTIME_SRCS) of the full Exec
library. They were placed in runtime_filter_exec_primitive_test.cpp, compiled into
the ExecPrimitive-scoped exec_primitive_test binary, which does not link that
library -- so the link failed with an undefined symbol while the header-only
ExecPrimitive types still resolved.

Move the four merger cases (and their helpers) into a new
exec/pipeline/runtime_filter_merger_test.cpp compiled into starrocks_test, which
links the full pipeline library, and drop the now-unused runtime_filter_types.h
include from the ExecPrimitive test. No test logic changes.
